+++ mg_custom_nodes/KLL535_ComfyUI_Simple_Qwen3-VL-gguf_20260821/__init__.py
@@ -2,9 +2,10 @@
 #__init__.py
 
 from .qwen3vl_node import SimpleQwen3VL_GGUF_Node
-from .utils_node import MasterPromptLoader,SimpleStyleSelector,SimpleCameraSelector,UnloadQwenModel,SimpleRemoveThinkNode,SimpleTriggerNode,TextToBatchNode,SimpleTextInsertNode,SimpleTextReplaceNode,SimpleJoinStringsNode
+from .utils_node import MasterPromptLoader,SimpleStyleSelector,SimpleCameraSelector,UnloadQwenModel,SimpleRemoveThinkNode,SimpleTriggerNode,TextToBatchNode,SimpleTextInsertNode,SimpleTextReplaceNode,SimpleJoinStringsNode,SimpleGifMaker
 from .deprecated_node import Qwen3VL_GGUF_Node
 from .configurator import Qwen3VL_ModelConfig, Qwen3VL_SamplingConfig
+#from .configurator import Qwen3VL_AdvancedConfig
 from .ideogram4 import Ideogram4JsonPreviewOnImage, Ideogram4JsonSwapCoordinates
 from .video_fragment_loader import SimpleLoadVideoFragment
 
@@ -24,9 +25,11 @@ NODE_CLASS_MAPPINGS = {
     "SimpleTextInsertNode": SimpleTextInsertNode,
     "SimpleTextReplaceNode": SimpleTextReplaceNode,
     "SimpleJoinStringsNode": SimpleJoinStringsNode,
+    #"Qwen3VL_AdvancedConfig": Qwen3VL_AdvancedConfig,
 
     #video
     "SimpleLoadVideoFragment": SimpleLoadVideoFragment,
+    "SimpleGifMaker": SimpleGifMaker,
 
     #ideogram4
     "Ideogram4JsonPreviewOnImage": Ideogram4JsonPreviewOnImage,
@@ -50,9 +53,11 @@ NODE_DISPLAY_NAME_MAPPINGS = {
     "SimpleTextInsertNode": "Simple Text Insert",
     "SimpleTextReplaceNode": "Simple Text Replace",
     "SimpleJoinStringsNode": "Simple Join Strings",
+    #"Qwen3VL_AdvancedConfig": "LLM Config (Advanced)",
 
     #video
     "SimpleLoadVideoFragment": "📸 Load Video Fragment",
+    "SimpleGifMaker": "📸 Simple Gif Maker",
 
     #ideogram4
     "Ideogram4JsonPreviewOnImage": "📐 Ideogram 4 JSON Preview",